Adds texture init for game_uis
