+++ bonus/srcs_bonus/game_bonus/t_game_destroy_bonus.c
@@ -26,6 +26,7 @@ void	t_game_destroy(t_game *game)
 		return ;
 	index = 0;
 	destroy_all_ui(game);
+	destroy_game_uis(game, &game->uis);
 	while (index < MAP_NB_IDS)
 	{
 		animation_textures = game->anim[index].textures;
