+++ bonus/srcs_bonus/mlx_api_bonus/t_image_utils_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   t_image_utils_bonus.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
+/*   By: codekiwi <codekiwi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 12:03:23 by mhotting          #+#    #+#             */
-/*   Updated: 2024/08/27 14:11:57 by brappo           ###   ########.fr       */
+/*   Updated: 2024/09/05 00:26:50 by codekiwi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
